requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
fastapi
python-dotenv